Add cast in async mysql
Summary: This is passed to a function which takes a `Variant`, and so
hits an implicit construtor. There are a lot of overloads on the
constructor; on Linux, this `size_t` hits one. But on OS X, where
`size_t` is `long unsigned int`, we're missing an overload. Since this
is the single place in HHVM we need that overload, just cast it to
`uint64_t`.

Reviewed By: @paulbiss

Differential Revision: D2059317