Preserve __thread variables in pthread exit handler on Mac.
Summary: The __thread variables in class T will be freed when pthread calls
the destructor function on Mac. We can register a handler in
pthread_t->__cleanup_stack similar to pthread_cleanup_push(). The handler
will be called earlier so the __thread variables will still exist in the
handler when the thread exits.

Part of #4444.
Closes https://github.com/facebook/hhvm/pull/5171

Reviewed By: @sgolemon

Differential Revision: D1991022