Fix DynamicContentCache and static content under FastCGI
Summary: When sending static content via sendStaticContent(), for example on a
DynamicContentCache hit, Transport::onSendEnd() was not called, and so
the webserver would hang forever waiting for an FCGI_END_REQUEST record.

Also, the key used for DynamicContentCache was different when storing
and retrieving, so it always missed. It used reqURI.path() when
retrieving, and reqURI.absolutePath() when storing, which may have
possibly been the same under certain test circumstances but is not the
same under FastCGI.
Closes https://github.com/facebook/hhvm/pull/5310

Reviewed By: @sgolemon

Differential Revision: D2056781