remove dead code update_pos
Summary: A while back we had this function that was updating positions instead of redeclaring.
This was in a pre-sharedmem world where data-nodes could actually run code.
I don't think it makes much sense today anymore, in fact, I am not sure it would be a win to
run a substitution vs simply redeclaring the types.
It can always be revived if we really need it.

Reviewed By: @int3

Differential Revision: D2062236