Add libunwind support for JIT.
Summary: 'z' field is required in the augmentation string to indicate length. Several
other required fields in libunwind are added. __register_frame needs the FDE
address, instead of the CIE address. _Unwind_GetIP is needed to get rip instead
of _Unwind_GetGR with Debug::RIP, because the rip number is -1 in libunwind
instead of 16. This change is compatible with both gcc and clang.

The cfi directives are enabled on Mac because they compiles with clang. All
quick tests pass in JIT mode after this change, expect 2 unrelated failures.
Part of #4444.
Closes https://github.com/facebook/hhvm/pull/5327

Reviewed By: @sgolemon

Differential Revision: D2061994