Drive-by cleanup of SrcKey
Summary: Need to add a field to SrcKey; might as well make it pretty.

This changes the assert in the constructors to check that the top bit is
not set, rather than to cast to a signed int and check for positivity,
because we're going to hijack another Offset bit soon.

Reviewed By: @ottoni

Differential Revision: D2046897