Fix various issues with gcc-4.9 build
Summary: This fixes the last fbconfig issue by updating to a rebuilt js_of_ocaml.

In order to successfully build hhvm, you have to uncomment the libiconv line in
hphp/util/TARGETS. I can't make that the default yet, because it breaks PHP5
compatibility (by fixing a glibc iconv bug in ucs-2 conversion).

This also fixes a couple of order of initialization issues that manifested with
gcc-4.9

Reviewed By: @edwinsmith

Differential Revision: D2064488