Fix ExtraArg leak when an VarEnv used in the same ActRec
Summary: Within HHVM, a memory leak occurs when extra arguments are passed to a function and that same function utilizes "variable variables". The root cause is that the variable variable usage causes the creation of a VarEnv, which takes ownership of the existing ExtraArg (created when the function is called with extra arguments) in the ActRec. However, even though the VarEnv has taken ownership, it never actually deallocates the ExtraArg. The solution is to add the logic to deallocate the ExtraArg at the proper point.

The erroneous behavior can be seen easily with the stripped-down reproducer P19833872. The object's destructor should be run when the function exits, but instead is run when the script completes. This reproducer has been added as a test-case.

Reviewed By: @jdelong

Differential Revision: D2061346