Fix generation of ConcatN opcode
Summary: The way that `srcStack` was being tracked was incorrect. In the
event of an opcode that does both pushes and pops, it might misattribute
the source to a previous opcode, since it was just based on the stack
size delta. Notably, `Concat` is such an opcode, meaning that
`ConcatPeephole` never did anything, since no value on the stack could
ever be attributed to a `Concat` opcode! (It's unclear how this ever
worked, the original diff looks broken since day 1?)

Instead, explicitly track the pushes and pops.

Note that we still don't generate `ConcatN` for a lot of OSS code,
including the very hot places in Drupal 7 that were the motivation here.
HHBBC will only emit this if it statically knows all the operands are
strings. Going to remove that restriction next.

Ref. https://github.com/facebook/hhvm/issues/5304

Reviewed By: @mxw

Differential Revision: D2065606