Remove some unused cfg.cpp loop functions, and replace findBackEdges
Summary: There seem to be multiple definitions of "back edge" in active use,
but I'll be adding a module that uses the definition "an edge who's target
dominates its source", so I'm renaming this one to avoid confusion (might need
to rename something in refcount-opts too).  Some books seem to call the weaker
concept "retreating edges", so use that.  Muchnick uses both definitions of
"back edge" in different places.  Dragon book seemed consistently on the
dominator definition.

Also: it looks like BackEdgeVisitor's algorithm was probably surprising about
the edges it could return when there is irreducible control flow (but this only
affects pretty printing so it's ok).  I /think/ it is not incorrect (for at
least one definition I've seen of "back edge"), but it looks like it could
identify edges that we probably want to think of as "forward edges" (because
they don't break the ability to visit nodes after predecessors in rpo).  E.g.:

       A
      / \
     B   C
    ^ \ /
    |  D<-+
    |  |  |
    |  V  |
    +--E--+

I think if it traverses down in the order A, C, D, E, then E->D will be
identified as a back edge, but E->B will be traversed because E is not in
m_path.  Then while looking at B, it will consider B->D, D will be in m_path,
and B->D will be added to the EdgeSet.  (For this new "retreating edges", we
want to return E->B and E->D, and for "back edges" (in the loops module) we
will only want E->D.)

Reviewed By: @swtaarrs

Differential Revision: D2071285