Don't ignore I<A> when declaring "extends I<A>, I<B>"
Summary: Problem:

class_type.tc_ancestors et al. are maps keyed by ancestor class name. If the same ancestor is listed multiple time, we would overwrite one with the other, depending on ordering, despite the other one possibly being needed to satisfy some constraints later. We should pick "more specific" one here, meaning the one lower in the class hierarchy.

The situation is straightforward(-ish) when A and B are in subtyping relation. I've struggled for a while with the case "what if they are not in common class hierarchy", i.e we have:

  interface I<+T> {}
  class A {}
  class B {}
  interface IAB extends I<A>, I<B> {}

Which one should we pick? I came to a conclusion that it doesn't matter, because the moment T is used anywhere, this declaration becomes invalid (see extends_multiple1.php) - forcing the developer to also explicitly implement I<CommonDescendant(A, B)>, and the choice of I<A> or I<B> doesn't matter, since it will be overwritten by this third class.

I suspect this might still be wrong, if the relation between A and B depends on some type variables and is not established yet when we do this check (not sure if it's possible). It will be better than what we do now anyway.

Reviewed By: @int3

Differential Revision: D2067091