Use GADTs to type the server-client RPCs
Summary: GADTs allow us to type the expected return value of a message, enabling
type-safe RPCs. This will make future refactorings easier and safer.

Not all of our server-client communication can be modeled via RPCs,
however; e.g. the build step streams back data continuously. Thus I've
categorized the messages into RPCs and streamed responses.

I also ended up cleaning up a bunch of redundant code. There is no
reason for STATUS to have so many different possible server responses;
directory mismatches are basically impossible since the directory name
is encoded into the socket name, and SERVER_DYING and PONG never
actually occur in practice.

hackProgram itself also has a bunch of useless functions. I think they
were previously required for the ServerFunctors interface, but somewhere
along the way the functor requirements changed.

During this rewrite, it also became apparent that a queue of asynchronously-
executed events would be useful. In particular, it is useful for delaying slow,
non-critical actions like logging, and it also allows us to treat `KILL` like an
ordinary RPC -- we put the `die_nicely` function on the async queue, ensuring
that we don't die before a response is sent to the client.

One unfortunate side-effect of this change is that the `client` target
now depends on `server_functors`. I had to put `ServerRpc` under
`server_functors` because it calls out to `PrologMain`, which in turn
depends on `server`, and thus `ServerRpc` cannot be in `server` because
that would create a circular dependency.

Reviewed By: @dlreeves

Differential Revision: D2059968