Don't optimize out concat binops with potential side-effects
Summary: Concat binops whose results aren't used can be optimized away by hphpc. However, this optimization is potentially faulty since if either of the operands are objects, their __toString() method will be invoked and cause arbitrary side-effects.

This can be seen easily with the reproducer P19837204. Without optimization, the echo inside the object's __toString() method is properly called, but with optimization it is not.

The solution is to mark any concat binop with either operand non-scalar as potentially having arbitrary side-effects. This will prevent the optimizer from removing the seemingly "dead" expression. This is deliberately conservative because the lack of type information at this stage prevents us from limiting this to only object operands.

Also added a test case to exercise this behavior.

Reviewed By: @markw65

Differential Revision: D2069031