Fix IRGS's profTransID for inlined blocks
Summary: Blocks that are created due to inlining don't have a valid
ProfTransID, so don't update IRGS.profTransID for them.  This will
just preserve the previous value, which is the proper value given that
inlined blocks must be translated right after the initial block for
the profiling translation containing them.  For non-PGO (Live)
translations, this will just keep IRGS.profTransID as kInvalidTransID.

While here, make the assert in getTransId() stronger, and also rename
getTransId and hasTransId to getTransID and hasTransID to better match
the style guide.

Reviewed By: @jdelong

Differential Revision: D2071102