Introduce ARef class to load elimination
Summary: Create an alias class for RefData referenced by SSATmps, and update memory
effects and load elimination to avoid redundant loads/checks of the inner ref.

This diff also includes an improvement to the union operator of alias classes.
When two different specializations are allowed, we try to keep the most useful
one, instead of discarding all the specialization. Since the ranking of
'usefulness' is fixed, we still get communitivity. (A similar idea might be
useful for Types too.)

Reviewed By: @jdelong

Differential Revision: D2069474