Fix an insert-asserts.cpp issue for critical edges
Summary: I fixed a bug for critical edges in insertRefCountAsserts, but added
one for insertStkAsserts.  The second case seems like it should be "ok" to
insert at a critical edge---and at first I wanted to just let it potentially
insert the assert "too many times".  But this isn't actually ok: we don't know
if every incoming edge made a Call, so the assert is too strong.  (If we care
enough at some point we could split this edge to insert the assert, but for now
just make the code safe.)

Reviewed By: @markw65

Differential Revision: D2078757