Make IR invariants hold after irgen; check an unchecked one; fix disabling DCE
Summary: One of the invariants is that no blocks have predecessors that aren't
reachable from the entry block.  But right now this isn't actually true until
we've run dce at least once, check.cpp wasn't asserting it, and we also never
run check.cpp before we've done at least one full dce.  It's also problematic
that before DCE, it was not clear which IR invariants are required to hold
(because it's clearly not all of them), and moreover, turning off the DCE
runtime option would allow this invariant to go unfixed between passes.

So, this diff splits DCE into two kinds: mandatoryDCE and fullDCE.  The
contract is that after running a pass which may cause violations of invariant
5, you must at least run manditoryDCE, but may run fullDCE instead.  Both of
those are guaranteed to restore the invariant, and may assume all other IR
invariants hold.  The rest of fullDCE is conditional on the runtime option.

Before settling on this I looked into how hard it would be to just make irgen
keep these invariants automatically: all the control flow helpers get a bit
more complicated than I want, and it's also pretty hard to reason about
when/whether gen time simplification can lead to violations of it (basically it
can), so it seems easiest to do it this way for now.  Also, relaxGuards could
leave this invariant unfixed before a simplify() run, and that could leave it
unfixed before running clean.

Another possible solution to this is to remove the invariant that there are no
unreachable blocks.  This might be better because it means DCE is just a pass
like any other pass, with the same contract as anything else.  But it would
also means you can't just leave it up to DCE to fix those, and more passes
might start with a redundant if (removeUnreachable()) reflowTypes.  So at least
for now I think this might be the way to go.

Reviewed By: @ottoni

Differential Revision: D2078795