Don't mimic HHBC guard relaxation in FPassL
Summary: We have giant retranslation chains in array_filter---it looks like it
might be because we're doing an FPassL to get the values into the callee.  We
can't change these FPassLs to FPassC or whatnot because they are created by a
WIter instruction (and have unknown reffiness statically).

Maybe this will help.

Reviewed By: @ottoni

Differential Revision: D2082419