Simple script for OSS repo mode
Summary: Enabling repo mode requires some slight shenanigans, which this
script can manage for folks. A followup diff to the packaging repo will
install this as `/usr/bin/hhvm-repo-mode` in our deb packages. I'm not
sure how portable this is beyond Debian/Ubuntu; most of it is general,
but the INI path and the `service` calls may not be; I wasn't sure how
to do that portably.

Ref. https://github.com/facebook/hhvm/issues/5356

Reviewed By: @paulbiss

Differential Revision: D2083261