Revert "Add skip reporting to hphp/test/run tests."
Summary: This reverts commit 35679bd97cf26013fec852ffd69bc89ba6c4ceda.

We skip over 2000 tests per build, and run many builds per diff. Reporting
all the skips is completely pointless, and just adds noise to the diff
(plus, even with this diff, we only reported 300 or so skips, so either
phabricator is intentionally limiting the number reported, or the diff
to enable skip reporting is broken in some way).

Reviewed By: @fredemmott

Differential Revision: D2087730