Fast Object <=> Array cast
Summary: Two major changes in this diff:
 - Fix get_object_vars to return an array that retains refiness
 - Allow sharing of dynamic property arrays (in particular when objects are
   created from casts on arrays or by clone).

To support sharing, dynamic property arrays are copied during set operations and
eagerly when non-const TypedValue pointer are extracted (so that if we write to
them the change does not effect anyone sharing the array).

With these changes and some optimizations to toArray and toIterArray enabled
by the new array sharing I'm seeing a 2% speedup for wordpress (it does a lot
of these casts).

Reviewed By: @jdelong

Differential Revision: D2072576