Add an option to enable hotcfg/wholecfg regions only for functions marked as hot
Summary: This is to try to reduce the overhead of JITing larger regions.  When this new option is enabled, for functions that are not marked as hot, we'll bail to the hottrace region selector when the new runtime option is true.

I'm leaving the new option disabled for now, since I couldn't measure any impact in my tests.

Reviewed By: @markw65

Differential Revision: D2087042