Reduce checks in Implode
Summary: We had a few unncessary checks in `Implode` for the common cases. This diff reduces them. Don't check
for containerish-ness, don't keep checking whether we are at the first string so we don't put a
delimiter, and don't keep checking if the string length is greater than 0 (it will be so rare that,
a once in a while all to memcpy with 0 bytes is ok).

Small percentage wins on Drupal and Mediawiki, relatively flat on Wordpress.

Reviewed By: @fredemmott

Differential Revision: D2095161