Get rtt out of NormalizedInstruction::inputs
Summary: We've wanted to do this for a while, since it's not always clear
exactly what the type means or where/when it came from. The main motivation for
doing it now is that it will greatly simplify how the new bytecode-based type
predictions interact with minstrs, which used to read some input types from the
NI. I started by just deleting the rtt member in DynLocation and fixing up any
usage sites to get the type from the canonical location, then it fanned out
into a bunch of cleanup/refactoring that was necessary to keep things working
(or trivial to do now):

- The only other DynLocation member was a Location, so DynLocation has been
  deleted.
- The Type argument has been removed from the top/pop functions in
  irgen-internal.h. The only types we ever passed in were what we statically
  knew from stack flavors, and we assert those before translating instructions
  anyway.
- As a consequence of the previous item, extendStack() should never create
  values of type StackElem anymore.
- getPropertyOffset() now lives in irgen-minstr.cpp, the only place that needs
  it.
- getPropertyOffset() and getCurrentPropertyOffset() have been simplified a lot
  due to the death of rtt, and no longer modify their Class* arguments.
- A few helpers related to guard relaxation in irgen-minstr.cpp now take IRGS&
  instead of MTS&. This is a step towards being able to call them before we've
  entered Mr. Translator and created an MTS, which is how we'll determine which
  type prediction to use.
- needFirstRatchet() was inspecting the base value input without first
  verifying that it was safe to do so.
- Delete numberStackInputs() in irgen-minstr.cpp. It hasn't been needed since
  analyze() died and the stack offsets in Location became predictable.
- Inline locPhysicalOffset into its single remaining caller.
- Fix a couple incorrect comments in type-specialization.h.
- Pull lambda from emitInputChecks() into its own function.
- Turn off HHIRRelaxGuards by default if HHIRConstrictGuards is on.

The better stack types can significantly reduce code size on certain exit
paths, since we'll now use the proper tracked type (rather than Gen) for
loading/storing/DecRefs. It doesn't affect the main code path, where
preOptimizeLdStk was already using the tracked type. We should look at turning
gen-time optimizations on for non-main blocks now that we track their state but
that can be a separate diff.

Reviewed By: @jdelong

Differential Revision: D2087314