Fix assertions in store-elim
Summary: When resolving cycles of phis, its possible that
one cycle depends on another, and that after resolving the
second one, the first only has a single input.

There's also a case where some of the inputs to a cyclic phi are
in progress when we try to resolve it, so handle that too.

Reviewed By: @jdelong

Differential Revision: D2101344