Count IntervalTimer samples between surprise checks
Summary: If we're in an operation that consumes a lot of wall time without
surprise checks, we can potentially miss IntervalTimer samples.  Fix this by
counting the number of times the timer expired since the last surprise check
and passing it to the callback function to use as needed.

Also fix a race condition (in the same piece of code) that could cause us to
miss samples entirely if the timer happens to go off at the same time as the
check in RunCallbacks.

Reviewed By: @billf

Differential Revision: D2100872