Leave the TC an x64 ret when returning control from a resumed async function
Summary: The return IP in the ActRec is always uniqueStubs.callToExit, which
adds 8 to rsp (to skip over the return ip to enterTCHelper) and then jumps
there.  So if we just ret from the TC we avoid mispredicting the return and
don't need the stub.

Note: the llvm side of this doesn't get the optimization right now, because we
need it to execute a return to the actual return address that's on the top of
the native stack.  @maksfb: is that possible with your EvalJitLLVMRetOpt code?

Reviewed By: @ottoni

Differential Revision: D2106195