Use empty "other" when iterating over configuration
Summary: Pass an empty hdf/ini when iterating over the other style config.
This will fix will be more evident with the `VirtualHost`, etc.
conversions to ini, but I wanted to get this fix out there now so
it can be used in those diffs after we land the big `Variant`
conversion diff.

Reviewed By: @fredemmott

Differential Revision: D2106336