Decouple hh_format from Relative_path
Summary: It doesn't make sense for hh_format to use relative paths,
since it doesn't need to be limited to operating under some root. It
also doesn't make sense to require `--root .` and / or a `.hhconfig`
to run hh_format.

The only time the root comes in useful is when we process a diff /
patch, and that can be treated independently of the Relative_path stuff.

Thus, this commit fixes things so that `--root` is only useful in diff
mode.

I've replaced the uses of Relative_path.t by Path.t. I could have replaced
them by `string`s instead, but this seems more principled. I also
discovered private types recently and it seemed like a good fit for
Path, so here it is... I can use normal abstract types if there are any
objections

  cd format
  grep Relative_path # only usage is to wrap the Path.t so that we can pass it to Parser_hack

Reviewed By: @pikatchu

Differential Revision: D2105938