Fix a load-elim bug when simplify() folds a CheckType into a Jmp
Summary: It would replace a sequence like:

   t1 :: Obj
   t2 = CheckType<BoxedCell> t1 -> B

with

   t1 :: Obj
   t2 = Mov t1
   Jmp B

This means that uses of t2 in the fallthrough block would now have a t2 of type
Obj, which isn't a subtype of BoxedCell, which can break various things.  (In
my case, load elim was freaking out because memory_effects claimed there was an
ARef location in an object, when it visited the LdRef in the now-unreachable
block.)

Instead, it turns out in this case we don't need to keep t2 defined, if the
simplifier is running after gen time.  At gen time, we can return a Bottom
because the dest of the check type irgen was trying to create is unreachable.
It also turns out we need to prevent load-elim from visiting blocks that the
simplifier made unreachable.  Discussion about why is in the code.

Reviewed By: @ottoni

Differential Revision: D2108705