Fix some bugs with AliasAnalysis for stacks (bring back the stk_expand_map)
Summary: In D2080432 stack_ranges was removed, but it wasn't exactly the same
as stk_expand_map: you also only had an entry in the stack_ranges map if every
location in the range was assigned a bit.  This was causing store-elim to
sometimes eliminate stack stores for PureSpillFrames when it ran out of aloc
bits.  We could probably store this information in a way that doesn't involve
two hashtables, but for now just move it a little more like how it was (and
keep the rest of that diff).

Reviewed By: @binliu19

Differential Revision: D2109977