memcpy rewriting
Summary: Reimplement memcpy for x86_64 platform.

The code is carefully laid out to expose parallelism, be cache friendly, and to avoid storing a jump table.

Explicit jccs seem better than indirect jmps. Thanks @ottoni for the suggestion.

However, this implementation is not optimized for maximizing the bandwidth for large copies.  We may need AVX for that.

A further idea is to make a specialized version for StringData. That could be done separately.

Reviewed By: @ottoni

Differential Revision: D1832458