Add ExitPlaceholders
Summary: If we want to hoist CheckTypes, we need to have already
generated exit blocks at the point where we want to hoist a particular
CheckType. ExitPlaceholder is a new IR node that serves this purpose.
They are inserted at IR gen time with an appropriate exit block and then
removed later in the compilation pipeline after any phases that care
have used them.

Reviewed By: @jdelong

Differential Revision: D1784788