Peephole optimize away trivial copy2 sequences.
Summary: Sometimes we CodeGen copy2{a,b,c,d}, (from an AssertType) and it
register-allocates to an xchg(r1,r2). And because of nearby phi
instructions, we end up inserting xchg(r2,r1) before it, resulting in
this clowntown:
  xchg rax,rcx
  xchg rax,rcx

This diff adds a simple peephole pattern to remove pairs of copy2
instructions that cancel each other out.

Reviewed By: @ottoni

Differential Revision: D2110676