Fix some problems with gvn on array instructions
Summary: The NameValueTableWrapper array doesn't have value semantics, so it's
a bit of a bummer but we can't do this stuff if we haven't proven an array
can't be $GLOBALS.  We should probably make that thing not be a KindOfArray,
and we might eventually also be able to assume if an SSATmp has an arrSpec with
a RAT array in it, then it can't be a $GLOBALS pointer.  But for now just fix
the problem.

Reviewed By: @swtaarrs

Differential Revision: D2111789