`ConstructInstance` returns an exact type
Summary: We already figure this out if there aren't any subtypes of the
constructed class, but it should be generally true that `new foo()`
returns a `foo`. Technically extension classes can do whatever they
want, but due to the "we already figure this out" that would be broken
anyways.

D1469306 changed `NewInstance` to do this nearly a year ago, if you want
to look at lineage.

Reviewed By: @alexmalyshev

Differential Revision: D2076538