Add ability to list tests and run a specific test to hphp framework suite.
Summary: Add an option to list all tests given a list of frameworks. Call
the framework runner as normal but add --list-tests as an argument (before
the list of frameworks itself due to limitations in the argument parser).
Similarly, add an option to run a single test. Call the framework runner as
normal but add --run-single-test <test> where <test> is one of the tests
returned by --list-tests.

Reviewed By: @paulbiss

Differential Revision: D2115193