Start on Windows support for hh_client and family. This also adds support for compiling the .c files as C++, due to MSVC's C mode being terrible. This also implements the filesystem watcher, which is mostly adapted from the OSX implementation. These ports are targetted towards MSVC 2013.
This doesn't touch the makefile, and I've left out the support for the shared heap, due to my implementation of it known to be broken.
This does not make it possible to build on windows, as I haven't touched the build system in this commit. OCamlbuild on Windows does not play well with the C files (it's possible, but not practical).