Improvements to serzialize_memoize_param
Summary: The key thing about this function is that different objects can't create colliding keys. Other than that, we can use whatever we want as a key as long as it's consistent. This makes a couple tweaks:
1: If a string starts with a letter, we just use it as is. Because it starts with a letter we know that it won't collide with integers used as keys, and also won't collide with anything produced by fb_compact_serialize, because output from that always starts with a charachter less than FB_CS_MAX_CODE, which is static asserted to less than '0'. It also won't collide with any of our static strings we use as keys since I modified them to start with charachters between FB_CS_MAX_CODE and '0'.
2: Gives a static string empty collections as well empty arrays. This is safe as we handle arrays and collections as the same thing in fb_compact_serialize as well (and so did DynamicYield).

In both of these cases, by calling fb_compact_serialize less, we avoid string allocs and should save some CPU. This function is .35% GCPU and don't expect more than a .1 win. So I expect this to be in noise in perflab, but I think it's strictly better

Reviewed By: @bertmaher

Differential Revision: D2112543