Selectively include DV inits and function bodies in the same region
Summary: When regenerating the function prologues along with their
corresponding DV funclets, the wholecfg selector was going into the
function body, effectively duplicating the whole function for each DV
entry.  Doing that can cause too much code duplication, but it can
also allow stores and type checks to be eliminated.  So this diff
selectively decides when to do that. DV inits are now JITed separately
if the function has multiple proflogues and the size of the function
is above a certain threshold.

The PGO region selectors already stop when they reach a block with a
SrcKey that is the start of a region that has already been created
(those SrcKeys are marked as "optimized").  So this diff leverages
that check and modifies regeneratePrologues in order to get the
desired effect.

Reviewed By: @mxw

Differential Revision: D1956841