Remove exit placeholders before store/load elim and other passes
Summary: I believe doing it after store-elim is probably disabling a lot of
SpillFrame removal for inlined functions.  (But I haven't verified that all
inlined functions have an ExitPlaceholder inside of them, so it may not be as
bad as that .... the ones that do will never have their SpillFrame removed
though.)  This will need to be revisited when someone looks at working on LICM
again, because it wants ExitPlaceholders if it's going to hoist loop invariant
guards.

Reviewed By: @mhahnenberg

Differential Revision: D2120546