Make StringData::Make methods initialize the ref-count to one
Summary: The family of StringData::Make methods initialize the ref-count of the
new string to zero. However, this is wasteful, because in all cases the
ref-count is immediately incremented either explicitly, or implicitly via the
String or Variant constructors.

To avoid this inefficiency, have the StringData::Make methods initialize the new
string's ref-count to one. This requires a bit of code churn, as all the places
that were passing the StringData into String or Variant now need to use the
attach methods (to avoid double-incrementing). However, a large number of
needless increments can be reduced.

This conversion also caught a couple of probable memory leaks.

Reviewed By: @edwinsmith

Differential Revision: D2118487