Deconflate preconditions and predictions
Summary: Currently the tracelet region selector and the region
translator conflate predictions and preconditions. PGO uses
preconditions to guide which blocks it stitches together. The tracelet
region selector uses predictions to seed which types we might want to
check for based on the types of live values.

This diff separates the two of these notions in RegionDesc::Block. We
now have separate maps for each.

Reviewed By: @jdelong

Differential Revision: D2094433