Fix a bug in regionizeFunc causing some arcs to not be marked as covered
Summary: If the region had an arc from a block B1 created due to inlining to
another block B2 from a different profiling translation, regionizeFunc
wasn't marking them as covered.  This would result in a potentially
unreachable Optimize translation to be created starting at block B2.
This diff adds a profiling TransID field to each block, and uses it to
mark such arcs as covered.

Reviewed By: @jdelong

Differential Revision: D2126006