Get rid of last frame creation site in MethodCache::handleSlowPath
Summary: We were calling a function with 7 args, so it had to set up a frame to
pass an argument on the stack. The 7th argument is just something loaded from
another argument, so redo that load in the callee rather than passing the
argument on the stack. There are now no non-tail calls in handleSlowPath(), gcc
can avoid setting up a frame in part of readMagicOrStatic(), and both functions
are smaller.

Reviewed By: @jdelong

Differential Revision: D2130142