Optimized FCallBuiltin for max with 2 operands
Summary: Emitter now generates an FCallBuiltin with max2 when there are two operands.
It contains a special case for int. The max2 C++ function is added to
ext_std_misc (in SystemLib). This is an attempt to bring back the 2 operand
optimization Owen had added to max earlier, which was lost when max was
converted to HNI last winter.

Reviewed By: @jdelong

Differential Revision: D2114083