Hoist fallbackccs above spills
Summary: In the cases vmsp isn't eliminated, it may get spilled and result in
bad code due to the decision of spilling values as soon as they're
defined.  More precisely, we'd start a translation with something like:

MAIN:
  lea rbx, ptr [rbp-0x60]
  lea rsp, ptr [rsp-0x10]
  mov qword ptr [rsp], rbx
  cmp byte ptr [rbp-0x8], 0x20
  jnz COLD
  ...

COLD:
  lea rsp, ptr [rsp+0x10]
  jmp NEXT_RETRANSLATION

When the guards fail, this code is bad for a couple of reasons: it
unnecessarily allocates and deallocates the spill space, and it causes
jumps to the cold code section by preventing the jcc in code.main from
being smashed to point to NEXT_TRANSLATION.

This diff adds a simple optimization that moves any spill code before
fallbackcc Vasm instructions to after the fallbackccs.  This pass is
called after spill code is inserted and before the instructions to
allocate and deallocate the spill space are inserted.  The result is
that the spill space ends up not being allocated (and deallocated)
until after the guards succeed.  For the example above, the generated
code now looks like:

MAIN:
  lea rbx, ptr [rbp-0x60]
  cmp byte ptr [rbp-0x8], 0x20
  jnz NEXT_RETRANSLATION
  lea rsp, ptr [rsp-0x10]
  mov qword ptr [rsp], rbx

Reviewed By: @edwinsmith

Differential Revision: D2126595