Allow anonymous functions to be subtypes of non-anonymous functions
Summary: Currently Typing_subtype does not do anything with our type for anonymous function Tanon so it always falls back to Typing_unify. Unfortunately this means declaring a return type for your anonymous function will lead to interesting Hack errors. For instance:

```
function error((function(): mixed) $x): void {}

function test(): void {
  $fun = (): int ==> 0;
  error($fun);
}
```

Will fail the type checker because ##int## is unified with ##mixed## instead of checking if ##int## is a subtype of ##mixed##. I add case in subtyping when the super type is a Tfun and the sub type is Tanon, to call the closure we have for the Tanon to get the return type and at least subtype that.

Reviewed By: @jwatzman

Differential Revision: D2133218