Allow disabling CallBuiltin optimization
Summary: CallBuiltin is great for speeding things up,
but it's terrible for code-coverage and profiling reports.

Allow disabling this optimization through:

  hhvm.enable_call_builtin=0

This surfaced an issue in error logging where we sometimes
double-report stack traces.

I also needed to fix some bits in xdebug and rerecord their
tests to reflect the improved data we get now.

Reviewed By: @paulbiss

Differential Revision: D2089156