Reuse PreFunc as first Func clone
Summary: Currently, PreClasses own a Func for each method, which is then cloned
by all Classes deriving from the PreClass.  This is unnecessary for the
first clone, which can just use the existing PreFunc, since it only
serves as a template---fields specific to cloned Funcs are reset after
they are cloned.

We add an atomic flag to Func which represents whether or not it is a
clone, and when we make the first clone, we just use the PreFunc instead
of allocating new memory.

Reviewed By: @markw65

Differential Revision: D2117197