Apply type conversion in NativeImpl stubs
Summary: This adds CastMem and CoerceMem instructions that mirror
CastStk and CoerceStk. I also removed the control flow from CoerceStk
and added it back explicitly at the ir level.

One issue that uncovered was that typecheck hoisting will happily convert
the sequence

     DefInlineFP
     InlineReturn
     CheckType -> B1
  B0:CallBuiltin
     ...

  B1:CoerceStk
     Jmp B0

In which we can optimize away the DefInlineFP and InlineReturn, into

     DefInlineFP
     CheckType -> side-exit
     InlineReturn
     CallBuiltin

In which we can't. In fact, I don't actually understand how its even a valid
transformation; there doesn't appear to be a correct bytecode to side-exit
to. So pending fixes for that, I'm disabling it by default.

Reviewed By: @ottoni

Differential Revision: D2149221