Generalize SMART_SIZES.
Summary: Generalize the formulation of SMART_SIZES so that it is possible to
generate kSmartIndex2Size (replaces s_index2size), as well as increasing
the range of size classes in the table so that additional fields can be
added to aid table-based generation of other data structures.

Remove LG_SMART_SIZES_PER_DOUBLING and hard-code
kLgSizeClassesPerDoubling to 2.  Although this diff removes SizeTable
(which couldn't handle any other value), in practice we are unlikely to
find other settings to be useful.

Reviewed By: @binliu19, @edwinsmith

Differential Revision: D2144708