Don't use libbfd in OSS build by default
Summary: libbfd is basically an interface into whatever the binutils
internals look like in the release that happens to be on your system.
This means there's no versioning, forwards, or backwards compatibility.
Basically, it's a PITA to use and package against. So we rip it out of
the OSS build, and ifdef out the little bit of code that was relying on
it.

This will lose us the stacktrace files in the OSS build, so we'll have
to rely more on folks attaching `gdb`, unfortunately. @paulbiss thinks
we might be able to use some folly functionality here eventually, but
bfd needs to go fairly soon.

sadpanda

Reviewed By: @paulbiss

Differential Revision: D2137703