Added SimpleXMLIterator
Summary: Added the SimpleXMLIterator as defined in standard PHP library (https://php.net/manual/en/class.simplexmliterator.php).
I had to make certain changes before i could implement this class. In hhvm SimpleXMLElement used to be an IteratorAggregate, while in the C implementation of PHP its a Traversable (but not an Iterator and not an IteratorAggregate!).
SimpleXMLIterator is a RecursiveIterator that extends from SimpleXMLelement. Because a class cannot be both a RecursiveIterator and an IteratorAggregate, i had do drop the IteratorAggregate interface from SimpleXMLelement and make it just a Traversable (just like in C PHP). As far as I know SimpleXMLElement is the only class in PHP that is a Traversable but not an IteratorAggregate or an Iterator, so a made a new kind-of API for it by creating a magic getIteratorForTraversable method. Not really sure if this was the right way to go, so plz let me know your thoughts...
Closes https://github.com/facebook/hhvm/pull/5352

Reviewed By: @JoelMarcey

Differential Revision: D2074073