Fix static constructor race in debug mode in memory manager
Summary: This assert may run when static constructors are being run, for
example when `StaticString`s are being constructed. But the memory
manager static constructor itself may not have run yet (in particular,
the `MemoryMnager::TlsWrapper` in `MemoryManagerInit()` may not have run
yet). But we don't actually care about this assert for static strings
according to @paulbiss, so we can just sidestep that.

The Mac build actually ran into this, FWIW.

Reviewed By: @paulbiss

Differential Revision: D2143834