Use a good, old version of perf to collect profile data for hfsort, and update linker_script
Summary: The latest version of perf is sometimes giving incorrect addresses.
This results in bad profiling data, and then messes up the the list of
hot functions produced by hfsort for the linker.  This diff adds a
working version of perf and changes the scripts to update the linker
script to use it instead.  This diff also includes an updated
linker_script.

Reviewed By: @binliu19, @markw65

Differential Revision: D2148747